deployment/cloud_function/requirements.txt

google-cloud-storage>=2.10.0
google-cloud-logging>=3.5.0
functions-framework>=3.4.0
requests>=2.31.0
